backend/requirements.txt
Flask>=2.0.1
psycopg2-binary>=2.8.6
Werkzeug>=2.0.1
